// projects/ui/src/lib/components/po-toaster/enum/po-toaster-type.enum.ts
/**
 * @usedBy PoToasterComponent
 *
 * @description
 *
 * Define os tipos possíveis para o `PoToasterComponent`.
 */
export enum PoToasterType {
  /** Tipo de toaster para mensagens de erro. */
  Error = 'error',

  /** Tipo de toaster para mensagens informativas. */
  Information = 'information',

  /** Tipo de toaster para mensagens de sucesso. */
  Success = 'success',

  /** Tipo de toaster para mensagens de atenção. */
  Warning = 'warning'
}